[ style ] 에러 모달 컴포넌트로 분리
