+++ src/@components/signUp/signupProfile.tsx
@@ -19,6 +19,7 @@ import { CategorySelectType } from "../../type/CategoryChecksType";
 import { EditDataType } from "../../type/editDataType";
 import { checkMaxInputLength } from "../../utils/uploadPage/maxLength";
 import useHover from "../../utils/hooks/useHover";
+import ProfileWarning from '../@common/profileWarning';
 
 export default function SignupProfile(props:SignupProfilePropsTye) {
   const {setStep, userProfile, setUserProfile}=props;
@@ -40,7 +41,6 @@ export default function SignupProfile(props:SignupProfilePropsTye) {
     FUNK: false,
   });
   const [contactInput, setContactInput]=useState<string>("");
-  const { hoverState, changeHoverState } = useHover();
 
   function getInputText(e: React.ChangeEvent<HTMLInputElement>) {
     setHashtagInput(e.target.value);
@@ -128,24 +128,7 @@ export default function SignupProfile(props:SignupProfilePropsTye) {
         <HashtagContainer>
           <HashIconWrapper>
           <ProfileEditHashtagIcon />
-          <WarningIcon onMouseEnter={(e) => changeHoverState(e)} onMouseLeave={(e) => changeHoverState(e)}>
-              {hoverState ? (
-                <>
-                  <HoverHashtagWarningIcon />
-                  <WarningTextWrapper>
-                    <WarningText>
-                      1. 해시태그는 최대 3개까지 추가 가능합니다.
-                      <br />
-                      2. 최대 10자까지 작성이 가능합니다.
-                      <br />
-                      3. 트랙의 분위기에 대해 설명해주세요. (ex. tropical, dynamic)
-                    </WarningText>
-                  </WarningTextWrapper>
-                </>
-              ) : (
-                <HashtagWarningIcon />
-              )}
-            </WarningIcon>
+          <ProfileWarning/>
             </HashIconWrapper>
           <InputHashtagWrapper>
           {hashtags.map((hashtag, index) => {
@@ -345,7 +328,7 @@ const DesciprtionInput = styled.textarea<{row:number}>`
   word-break: break-word;
   border: none;
   background-color: transparent;
-  margin-top: 3.3rem;
+  margin-top: 3rem;
   overflow: hidden;
 
   border-bottom: 0.1rem solid ${({ theme }) => theme.colors.gray3};
@@ -392,54 +375,12 @@ const ProfileEditDescriptionIcon=styled(ProfileEditDescriptionIc)`
   width: 12.6rem;
 `
 
-
 const DeleteHashtagIcon = styled(DeleteHashtagIc)`
   width: 2.8rem;
   
   margin-left: -1rem;
   cursor: pointer;
 `;
-
-const WarningTextWrapper = styled.div`
-  height: 12.5rem;
-  width: 47.2rem;
-
-  position: absolute;
-
-  margin-top: 1.4rem;
-  background: rgba(30, 32, 37, 0.7);
-  backdrop-filter: blur(3px);
-  border-radius: 5px;
-`;
-
-const WarningText = styled.div`
-  ${({ theme }) => theme.fonts.description};
-  color: ${({ theme }) => theme.colors.gray2};
-
-  margin: 1.9rem 1.8rem 0.4rem 2.9rem;
-`;
-
-const WarningIcon = styled.div`
-  height: 3rem;
-  margin-top: 0.7rem;
-  border-radius: 5rem;
-
-  cursor: pointer;
-`;
-
-const HoverHashtagWarningIcon=styled(HoverHashtagWarningIc)`
-  width: 4rem;
-  height: 4rem;
-  margin-left: 42.6rem;
-  margin-top: -1rem;
-`
-
-const HashtagWarningIcon=styled(HashtagWarningIc)`
-  width: 4rem;
-  height: 4rem;
-  margin-left: 42.6rem;
-  margin-top: -1rem;
-`
 
 const HashIconWrapper=styled.div`
   display: flex;
